fuzz: count duplicate outputs under real-world harnessing
paradoxically, the number of dupes seems to go _up_ with better chunk sizes.
this is because the inputs are not varied enough.  the harness is providing
us with large fuzzbuffers and then flipping bits past the start where it
would make any difference.  so most of its flips produce duplicates simply
because the generator is ony looking at the first few bytes, because it
errors out after looking at those first few.  therefore, for a more accurate
dupe count, we only include ones that made it past that parsing phase.